+++ share/pot/common.sh
@@ -873,5 +873,12 @@ pot-cmd()
 	fi
 	. "${_POT_INCLUDE}/${_cmd}.sh"
 	_func=pot-${_cmd}
-	$_func "$@"
+	case ${_cmd} in
+		create|clone|import)
+			lockf -k /tmp/pot-potnet-lock "$_func" "$@"
+			;;
+		*)
+			$_func "$@"
+			;;
+	esac
 }